requirements.txt
python-decouple>=3.1
requests>=2.19.1
